django/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
dill==0.3.9
Django==4.2.18
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
filelock==3.17.0
idna==3.10
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nlp==0.4.0
numpy==2.0.2
openpyxl==3.1.5
openpyxl-image-loader==1.0.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xlwt==1.3.0
xxhash==3.5.0
